Rename package according with component task
