+++ pom.xml
@@ -4,8 +4,8 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>com.globalhitss.claropay.cercademi</groupId>
-  <artifactId>commerceetlip</artifactId>
+  <groupId>com.globalhitss.claropay.cercademi.job</groupId>
+  <artifactId>ipgeolocation</artifactId>
   <version>0.0.1</version>
 
   <name>commerceetlip</name>
@@ -143,7 +143,7 @@
           <artifactId>exec-maven-plugin</artifactId>
           <version>1.6.0</version>
           <configuration>
-            <mainClass>com.globalhitss.claropay.cercademi.commerceetlip.App</mainClass>
+            <mainClass>com.globalhitss.claropay.cercademi.job.ipgeolocation.App</mainClass>
           </configuration>
         </plugin>
       </plugins>
