+++ src/main/java/com/globalhitss/claropay/cercademi/job/ipgeolocation/util/IPTools.java
@@ -1,4 +1,4 @@
-package com.globalhitss.claropay.cercademi.commerceetlip.util;
+package com.globalhitss.claropay.cercademi.job.ipgeolocation.util;
 
 import java.util.LinkedList;
 import java.util.List;
